+++ CHANGELOG.md
@@ -85,3 +85,4 @@
 - Started implementing fuel logic
 - Fuel flow working
 - Removed instant fuel consumption, as its highly inaccurate using MAF calculations.
+- Start made on integrating obd activity into the main project